#1109 and #1041 Work in progress.  Adding linux/cygwin/msvc support to test_daily.sh

+++ trunk/buildserver/test_daily.sh
@@ -26,33 +26,41 @@ if [ "$PLATFORM" == "" ]; then
     fi
 fi
 
+##
+# figure out today's build
 date=$(date '+%Y-%m-%d')
-
-echo date = $date
-echo url  = $JENKINS/$DAILY
-
 build=$(curl --silent $JENKINS/$DAILY/                 \
        |xmllint --pretty 1 - | grep $PLATFORM          \
        |grep $date | grep -v -e view | cut -d'"' -f 2  ) 2>/dev/null
+
+echo date  = $date
+echo url   = $JENKINS/$DAILY
 echo build = $build
-cd /tmp
-rm -rf $build
+
+##
+# collect build from server
+if [ -e /tmp/jenkins ]; then rm -rf /tmp/jenkins ; fi
+mkdir /tmp/jenkins
+cd /tmp/jenkins
 curl -O --silent $JENKINS/$DAILY/$build
 ls -alt $build
-if [ -e dist ]; then rm -rf dist ;fi
+if [ ! -e $build ]; then echo '*** $build has not been downloaded ***' ; exit 1; fi
 
 ##
 # expand the bundle
+if [ -e dist ]; then rm -rf dist ;fi
 tar xzf $build
 if [ ! -e dist ]; then echo '*** no dist directory ***' ; exit 1; fi
 
+##
 # enter the dist and test it
 cd dist
+grep_args="-e libexiv2 -e ^date -e ^bits -e ^version -e ^time"
 case $PLATFORM in
     macosx)
         # test the delivered exiv2
         DYLD_LIBRARY_PATH="$PWD/$PLATFORM/lib:$DYLD_LIBRARY_PATH"
-        $PLATFORM/bin/exiv2 -vV | grep -e libexiv2 -e ^date -e ^bits -e ^version -e ^time
+        $PLATFORM/bin/exiv2 -vV | grep $grep_args
 
         # compile, link and test the sample code
         echo ''
@@ -60,7 +68,49 @@ case $PLATFORM in
         ls -alt exifprint
         echo ''
 
-        exifprint --version     | grep -e libexiv2 -e ^date -e ^bits -e ^version -e ^time
+        exifprint --version     | grep $grep_args
+    ;;
+    
+    linux)
+        # test the delivered exiv2
+        LD_LIBRARY_PATH="$PWD/$PLATFORM/lib:$LD_LIBRARY_PATH"
+        $PLATFORM/bin/exiv2 -vV | grep $grep_args
+
+        # compile, link and test the sample code
+        echo ''
+        g++ -I$PLATFORM/include -L$PLATFORM/lib samples/exifprint.cpp -lexiv2 -o exifprint
+        ls -alt exifprint
+        echo ''
+
+        exifprint --version     | grep $grep_args
+    ;;    
+    
+    cygwin)
+        # test the delivered exiv2
+        PATH="$PWD/$PLATFORM/bin:$PATH"
+        $PLATFORM/bin/exiv2 -vV | grep $grep_args
+
+        # compile, link and test the sample code
+        echo ''
+        # cmd /c "vcvars 2013 64 && cl something"
+        # ls -alt exifprint.exe
+        # echo ''
+
+        exifprint --version     | grep $grep_args
+    ;;    
+
+    msvc)
+        # test the delivered exiv2
+        PATH="$PWD/$PLATFORM/bin:$PATH"
+        $PLATFORM/bin/exiv2 -vV | grep $grep_args
+
+        # compile, link and test the sample code
+        echo ''
+        # cp -R $PLATFORM/* /usr/local/
+        # g++ -I$PLATFORM/include -L$PLATFORM/lib samples/exifprint.cpp -lexiv2 -o exifprint
+        # ls -alt exifprint.exe
+        echo ''
+        ./exifprint.exe --version     | grep $grep_args
     ;;
 
     *) echo unknown platform $platform
